chore: move get default method in utils
